+++ bluetti_mqtt/bluetooth/__init__.py
@@ -3,13 +3,13 @@ import re
 from typing import Set
 from bleak import BleakScanner
 from bleak.backends.device import BLEDevice
-from bluetti_mqtt.core import BluettiDevice, AC200M, AC300, AC500, EP500, EP500P, EP600, EB3A
+from bluetti_mqtt.core import BluettiDevice, AC200M, AC300, AC500, AC60, EP500, EP500P, EP600, EB3A
 from .client import BluetoothClient
 from .exc import BadConnectionError, ModbusError, ParseError
 from .manager import MultiDeviceManager
 
 
-DEVICE_NAME_RE = re.compile(r'^(AC200M|AC300|AC500|EP500P|EP500|EP600|EB3A)(\d+)$')
+DEVICE_NAME_RE = re.compile(r'^(AC200M|AC300|AC500|AC60|EP500P|EP500|EP600|EB3A)(\d+)$')
 
 
 async def scan_devices():
@@ -31,6 +31,8 @@ def build_device(address: str, name: str):
         return AC300(address, match[2])
     if match[1] == 'AC500':
         return AC500(address, match[2])
+    if match[1] == 'AC60':
+        return AC60(address, match[2])
     if match[1] == 'EP500':
         return EP500(address, match[2])
     if match[1] == 'EP500P':