+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## FUTURE
 
+## 0.15.0
+
 * Add additional battery pack details for AC200M, AC300, EP500(P), and AC500
 * Add initial logging support for AC60 (and very basic polling)